Add CSS styles to all pages
Create a new file "styles.css" to act as an external CSS stylesheet for
index.html and all other pages.

Move location of "Home Page" link to bottom of the page, and center
images and captions.

Add background color to body and headings, and borders to headings and
lists.

Set default font to Times New Roman, with serif fallback.